Added google api calls and hard coded a JSON file of grocery items.
The simulation on iPhone does not get the actual lat and long er need
for the google api so we must find another solution. Also I hard coded
a json file of grocery items for now I think that is the best. I am
making a async call that is retrieving these results in the first view
controller almost instantly. Right now it will simply show them in an
NSLog but we will then push them through a segue on to the next view
controller so we show them using the accordions.